Use floats instead of int so we can still support 32 bit systems

+++ www/go/modules/community/maildomains/model/Domain.php
@@ -43,7 +43,7 @@ final class Domain extends AclOwnerEntity
 	/**
 	 * Default quota in bytes
 	 */
-	public int $defaultQuota = 0;
+	public float $defaultQuota = 0;
 
 	public string $transport;
 	public bool $backupMx = false;
